move initializePsychicApplication
