+++ boilerplate/api/src/conf/repl.ts
@@ -1,7 +1,7 @@
 import { loadRepl } from '@rvoh/dream'
 import * as repl from 'node:repl'
 import AppEnv from './AppEnv.js'
-import initializePsychicApplication from './initializePsychicApplication.js'
+import initializePsychicApplication from './system/initializePsychicApplication.js'
 import './loadEnv.js'
 
 const replServer = repl.start('> ')
