+++ boilerplate/api/src/conf/system/initializePsychicApplication.ts
@@ -1,6 +1,6 @@
 import { PsychicApplication, PsychicApplicationInitOptions } from '@rvoh/psychic'
-import psychicConf from './app.js'
-import dreamConf from './dream.js'
+import psychicConf from '../app.js'
+import dreamConf from '../dream.js'
 
 export default async function initializePsychicApplication(opts: PsychicApplicationInitOptions = {}) {
   return await PsychicApplication.init(psychicConf, dreamConf, opts)
